adding aruco marker example in
Adding simple aruco marker detector into the depth map and segment out where the aruco marker is located. Then all we are doing is to find the average in the mask. This is not accurate due to the hole in the point cloud, need to find an way to remove all the holes.